backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic-settings==2.6.0
httpx==0.27.2
python-dotenv==1.0.1
tenacity==9.0.0
psycopg2-binary==2.9.10
greenlet==3.1.1
